api/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
psycopg2-binary==2.9.10
celery==5.5.3
redis==6.4.0
apscheduler==3.11.0
pydantic==2.11.7
pyyaml==6.0.2
cryptography==45.0.6
